requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.0
matplotlib==3.9.3
seaborn==0.13.2
streamlit==1.41.1
python-dotenv==0.21.0
pydantic==2.8.2
pinecone-client==5.0.1
langchain==0.3.7
langchain_community==0.3.6
langchain-openai==0.2.8
langchain_core==0.3.18
langchain_pinecone==0.2.0
openai==1.54.0
torch==2.5.1
semantic_router==0.0.72
transformers==4.47.1
fpdf==1.7.2
fuzzywuzzy==0.18.0
httpx==0.27.2 # fix compatibility issues
sqlitecloud==0.0.83
